+++ filament/benchmark/benchmark_filament.cpp
@@ -21,7 +21,7 @@
 
 #include <filament/Box.h>
 #include <filament/Frustum.h>
-#include "details/Culler.h"
+#include "Culler.h"
 
 #include <utils/Allocator.h>
