+++ filament/src/FrameSkipper.h
@@ -17,7 +17,7 @@
 #ifndef TNT_FILAMENT_DETAILS_FRAMESKIPPER_H
 #define TNT_FILAMENT_DETAILS_FRAMESKIPPER_H
 
-#include "backend/Handle.h"
+#include <backend/Handle.h>
 
 #include <array>
