+++ filament/src/details/Engine.h
@@ -18,42 +18,42 @@
 #define TNT_FILAMENT_DETAILS_ENGINE_H
 
 #include "upcast.h"
+
+#include "Allocators.h"
 #include "PostProcessManager.h"
+#include "ResourceList.h"
 
 #include "components/CameraManager.h"
 #include "components/LightManager.h"
 #include "components/TransformManager.h"
 #include "components/RenderableManager.h"
 
-#include "details/Allocators.h"
 #include "details/BufferObject.h"
 #include "details/Camera.h"
+#include "details/ColorGrading.h"
 #include "details/DebugRegistry.h"
 #include "details/Fence.h"
 #include "details/IndexBuffer.h"
 #include "details/RenderTarget.h"
-#include "details/ResourceList.h"
-#include "details/ColorGrading.h"
-#include "details/Skybox.h"
 #include "details/SkinningBuffer.h"
+#include "details/Skybox.h"
 
-#include "private/backend/CommandStream.h"
 #include "private/backend/CommandBufferQueue.h"
+#include "private/backend/CommandStream.h"
 #include "private/backend/DriverApi.h"
 
 #include <private/filament/EngineEnums.h>
 #include <private/filament/UniformInterfaceBlock.h>
 
+#include <filament/ColorGrading.h>
 #include <filament/Engine.h>
-#include <filament/VertexBuffer.h>
 #include <filament/IndirectLight.h>
 #include <filament/Material.h>
 #include <filament/MaterialEnums.h>
-#include <filament/Texture.h>
-#include <filament/ColorGrading.h>
 #include <filament/Skybox.h>
-
 #include <filament/Stream.h>
+#include <filament/Texture.h>
+#include <filament/VertexBuffer.h>
 
 #if FILAMENT_ENABLE_MATDBG
 #include <matdbg/DebugServer.h>
